Decouple calculate fee from Wallet
